# fix: yield single buffers