--- a/src/encode.ts
+++ b/src/encode.ts
@@ -1,5 +1,6 @@
 import type { Source } from 'it-stream-types'
 import varint from 'varint'
+import { Uint8ArrayList } from 'uint8arraylist'
 import { allocUnsafe } from './alloc-unsafe.js'
 import { Message, MessageTypes } from './message-types.js'
 
@@ -17,7 +18,7 @@ class Encoder {
   /**
    * Encodes the given message and returns it and its header
    */
-  write (msg: Message): Uint8Array[] {
+  write (msg: Message, list: Uint8ArrayList): void {
     const pool = this._pool
     let offset = this._poolOffset
 
@@ -41,16 +42,11 @@ class Encoder {
       this._poolOffset = offset
     }
 
-    if ((msg.type === MessageTypes.NEW_STREAM || msg.type === MessageTypes.MESSAGE_INITIATOR || msg.type === MessageTypes.MESSAGE_RECEIVER) && msg.data != null) {
-      return [
-        header,
-        ...(msg.data instanceof Uint8Array ? [msg.data] : msg.data)
-      ]
-    }
+    list.append(header)
 
-    return [
-      header
-    ]
+    if ((msg.type === MessageTypes.NEW_STREAM || msg.type === MessageTypes.MESSAGE_INITIATOR || msg.type === MessageTypes.MESSAGE_RECEIVER) && msg.data != null) {
+      list.append(msg.data)
+    }
   }
 }
 
@@ -61,12 +57,16 @@ const encoder = new Encoder()
  */
 export async function * encode (source: Source<Message | Message[]>) {
   for await (const msg of source) {
+    const list = new Uint8ArrayList()
+
     if (Array.isArray(msg)) {
       for (const m of msg) {
-        yield * encoder.write(m)
+        encoder.write(m, list)
       }
     } else {
-      yield * encoder.write(msg)
+      encoder.write(msg, list)
     }
+
+    yield list.subarray()
   }
 }